[move-prover] make Verification variant taking a flavor enum
Existing implementation takes a static string with the assumption that
all verification variants can be statically known. This is true as of
now but even in this case, it feels better to use an enum instead of raw
strings to match the verification invariants.

This commit is also up to support an incoming transformation which may
create a statically unknown number of verification variants when
unrolling global invariants with generics instantiated.

Closes: #8745