[move-model] allow the spec model to be built when compiler has warnings
The model building seems to terminate at a state where an incomplete
`GlobalEnv` has warnings only. But the prover does not know that the
`GlobalEnv` is incomplete, so it produces no bytecode subsequently.

This commit allows the warning-only diagnostics to be added to the
`GlobalEnv` first and also allows the model building to continue from
there.

Closes: #8972
Closes: #10607