[move-prover] Renables invariant disabling analysis.
For some reason, before #8946 public functions which have invariants indirectly disabled have not been reported, and that PR disabled the logic for this to not block landing.

This PR renables the error reporting with better error messages for why invariants are disabled. With the help of this it then  fixes various instances in the framework where invariant disabling requires to introduce friend functions.

This also fixes visibility problems in move-lang/functional-tests and e2e-testsuite after we turned some functions into friends. Friend functions introduced in this PR are not visible to those tests. To mitigate, we introduce wrappers `foo_for_test` for each friend function `foo`, and call those from tests instead. Those functions should have a test_only attribute, however, this currently does not work because legacy unit tests and e2e tests do not see test-only functions. So we added the attribute but commented it out with a TODO.

Closes: #8988
Closes: #10627